% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblcompare.R
\name{udf_CreateChecksumDict}
\alias{udf_CreateChecksumDict}
\title{Get a dictionary of pairs of dependent tables and their identifier, for a given stored procedure}
\usage{
udf_CreateChecksumDict(server, db, proc)
}
\arguments{
\item{server}{Name of SQL Server Instance}

\item{db}{Name of database}

\item{proc}{Name of stored procedure}
}
\value{
a dataframe contains dependent tables and their unique identifier
}
\description{
Takes in the name of a stored procedure, database and the instance of a SQL Server,
    and return a dictionary of key values pairs of dependent tables and corresbonding identifier
}
