% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblcompare.R
\name{udf_GetChecksum}
\alias{udf_GetChecksum}
\title{Get a numerical unique identifier for a given table}
\usage{
udf_GetChecksum(server, db, tbl)
}
\arguments{
\item{server}{Name of SQL Server Instance}

\item{db}{Name of database}

\item{tbl}{Name of table}
}
\value{
value of the numerical unique identifer
}
\description{
Takes in name of a table, database and the instance of a SQL Server, and return an unique identifer based on data and schema
}
