% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblcompare.R
\name{udf_GetColumns}
\alias{udf_GetColumns}
\title{Get the columns for specified table}
\usage{
udf_GetColumns(server, db, tbl)
}
\value{
a vector contains columns of the table
}
\description{
Takes in name of a table and an instance of a SQL Server.
}
