% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblcompare.R
\name{udf_GetTableList}
\alias{udf_GetTableList}
\title{Get all the dependent tables for a stored procedure}
\usage{
udf_GetTableList(server, db, proc)
}
\arguments{
\item{server}{Name of SQL Server Instance}

\item{db}{Name of database}

\item{proc}{Name of stored procedure}
}
\value{
a vector contains all dependent tables
}
\description{
Takes in the name of a stored procedure, database and the instance of a SQL Server,
   and return the procedure's dependent tables
}
