% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblcompare.R
\name{udf_MergeSourceTarget}
\alias{udf_MergeSourceTarget}
\title{Merge the dictionary of source table and target table}
\usage{
udf_MergeSourceTarget(source_server, target_server, db, proc)
}
\arguments{
\item{source_server}{Name of source SQL Server Instance}

\item{target_server}{Name of target SQL Server Instance}

\item{db}{Name of database}

\item{proc}{Name of stored procedure}
}
\value{
a dataframe contains dependent tables, values of identifier on source server, and values on target server
}
\description{
Takes in a source proc information and target proc information, and return a merged result of their dictionaries
}
